+++ datamodel/include/adj_matrix_dense.h
@@ -1,26 +1,27 @@
 #ifndef _ADJ_MATRIX_DENSE_H_
 #define _ADJ_MATRIX_DENSE_H_
 
+#include "data.h"
 #include "adj_edges.h"
 
 class AdjMatrixDense
 {
 private:
-    long vertices;
-    long edges;
-    long** matrix;
+    int vertices;
+    int edges;
+    int** matrix;
 public:
-    AdjMatrixDense(long size);
-    AdjMatrixDense(long size, long* arr);
+    AdjMatrixDense(int size);
+    AdjMatrixDense(int size, int* arr);
     AdjMatrixDense(const AdjEdges& edges);
     ~AdjMatrixDense();
 
-    long num_vertices() const;
-    long num_edges() const;
-    void set_edges(long edges);
-    long size() const;
-    long* operator[](long index);
-    const long* operator[](long index) const;
+    int num_vertices() const;
+    int num_edges() const;
+    void set_edges(int edges);
+    int size() const;
+    int* operator[](int index);
+    const int* operator[](int index) const;
 };
 
 #endif
